+++ src/chain-wallet-libs/CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+## [0.5.0-pre7] - 2020-11-16
+
 #### wallet-js
 
 - Expose function to decode bech32 to byte array